finish rename in tests
